updated the csv file to go ahead and append information. Updated the README.md file. Organized the files on the app.
