+++ Fitness_App/README.md
@@ -1,0 +1,26 @@
+# My_Fitness_APP
+
+### Introduction 
+My_Fitness_App is an application which helps keep track of your calorie in take. The app provides a selection of foods and you have the ability to mark which foods you have ate in a day. After the selection is made the app would output the amount of calories you had intake and if you had met your goal or over exceeded.
+
+### Version Requirements
+The app requires:
+- Python version 3.12.4
+- Microsoft Visual Studio Code
+- Git
+
+### Installation
+Clone the repo: 
+
+- git clone (paste the https). Then write cd write_the_project_directory.
+
+Create Virtual Environment:
+- Hit (Ctrl+Shift+P) then type "Python: Create Environment" and select it then .venv then 3.12.4. Would then prompt you to install the requirements.txt file. 
+- if doesnt work follow these steps:
+    - Proceed to then write pip install -r requirements.txt (inside the command window).
+ 
+
+### Resources
+Food_List.py: 
+- The food_list.py is a useful default file. If there is no food file, we would run this file to generate a default of foods.
+    - A food file is a json file. The file intels the food name, the type, and the amount of calories the food consist of.
