--- README.md
@@ -1,2 +1,0 @@
-# My_Fitness_APP
-Daily Goal meetings
