--- omar/helpers/food_list.py
@@ -1,103 +1,0 @@
-import json
-
-foods = {
-    "Bugs" : {
-        "calories": 300,
-        "type": "Protein"
-    },
-    "Clementines": {
-        "calories": 100,
-        "type": "Fruit"
-    },
-    "Broccoli": {
-        "calories": 50,
-        "type": "Vegetables"
-    },
-    "Chicken": {
-        "calories": 285,
-        "type": "Protein"
-    },
-    "Pork": {
-        "calories": 275,
-        "type": "Protein"
-    },
-    "Salmon": {
-        "calories": 200,
-        "type": "Protein"    
-    },
-    "Eggs": {
-        "calories": 80,
-        "type": "Protein"
-    },
-    "Greek Yogurt": {
-        "calories": 280,
-        "type": "Protein"
-    },
-    "Apple": {
-        "calories": 100,
-        "type": "Fruit"
-    }, 
-    "Strawberries":{
-        "calories": 50,
-        "type": "Fruit"
-    },
-    "Pineapple": {
-        "calories": 85,
-        "type": "Fruit"
-    },
-    "Banana": {
-        "calories": 105,
-        "type": "Fruit"
-    },
-    "Avocado": {
-        "calories": 250,
-        "type": "Fruit"
-    }, 
-    "Carrot": {
-        "calories": 25,
-        "type": "Vegetables"
-    },
-    "Lettuce": {
-        "calories": 5,
-        "type": "Vegetables"
-    },
-    "Tumeric": {
-        "calories": 20,
-        "type": "Vegetables"
-    },
-    "Spinach": {
-        "calories": 20,
-        "type": "Vegetables"
-    },
-    "Kale": {
-        "calories": 10,
-        "type": "Vegetables"
-    },
-    "Milk": {
-        "calories": 120,
-        "type": "Protein"
-    },
-    "Corn Tortilla": {
-        "calories": 60,
-        "type": "Protein"
-    }, 
-    "Bell Peppers": {
-        "calories": 25,
-        "type": "Vegetables"
-    },
-    "Brussel Sprout": {
-        "calories": 40,
-        "type": "Vegetables"
-    },
-    "Blackberries": {
-        "calories": 60,
-        "type": "Fruit"
-    },
-    "Blueberry": {
-        "calories": 45,
-        "type": "Fruit"
-    }
-    }
-food = "foods.json"
-with open (food, "w") as file:
-    json.dump(foods, file, indent = 4)
