--- omar/output/Sat_Jan_1_2000.csv
@@ -1,4 +1,0 @@
-Name,Calories,Type
-Blueberry,45,Fruit
-Avocado,250,Fruit
-Strawberries,50,Fruit
