--- omar/output/Wed_Jan_26_2000.csv
@@ -1,14 +1,0 @@
-Name,Calories,Type
-Bugs,300,Protein
-Milk,120,Protein
-Banana,105,Fruit
-Pineapple,85,Fruit
-Strawberries,50,Fruit
-Apple,100,Fruit
-Lettuce,5,Vegetables
-Tumeric,20,Vegetables
-Greek Yogurt,280,Protein
-Eggs,80,Protein
-Salmon,200,Protein
-Pork,275,Protein
-Chicken,285,Protein
