--- omar/requirements.txt
@@ -1,1 +1,0 @@
-PySide6==6.7.2
